Move all code from Main.java to Duke.java
