+++ src/main/java/Duke.java
@@ -1,10 +1,54 @@
+import java.util.Scanner;
+
 public class Duke {
+    private static final String GREET_MESSAGE
+            = "Roarrr.... I'm Burning Charizard, tasked to burnnn down your tasks.\n" +
+            "Which task shall we burn today?";
+    private static final String BYE_MESSAGE = "Roarrr....Let's burn more tasks next time!";
+    private static final int BORDER_LENGTH = 70;
+
     public static void main(String[] args) {
+        greet();
+        Scanner sc = new Scanner(System.in);
+        System.out.println("What should Charizard do next?");
+        String input = sc.nextLine();
+        while (input.compareToIgnoreCase("bye") != 0) {
+            show(input);
+            System.out.println("What should Charizard do next?");
+            input = sc.nextLine();
+        }
+        exit();
+    }
+
+    public static void show(String message) {
+        drawBorder(BORDER_LENGTH);
+        System.out.println("    " + message);
+        drawBorder(BORDER_LENGTH);
+        System.out.println();
+    }
+
+    public static void exit() {
+        drawBorder(BORDER_LENGTH);
+        System.out.println(BYE_MESSAGE);
+        drawBorder(BORDER_LENGTH);
+    }
+
+    public static void greet() {
         String logo = " ____        _        \n"
                 + "|  _ \\ _   _| | _____ \n"
                 + "| | | | | | | |/ / _ \\\n"
                 + "| |_| | |_| |   <  __/\n"
                 + "|____/ \\__,_|_|\\_\\___|\n";
         System.out.println("Hello from\n" + logo);
+        drawBorder(BORDER_LENGTH);
+        System.out.println(GREET_MESSAGE);
+        drawBorder(BORDER_LENGTH);
+    }
+
+    private static void drawBorder(int length) {
+        for (int i = 0; i < length; i++) {
+            System.out.print("*");
+        };
+        System.out.println("");
     }
 }
